requirements.txt
bson==0.5.10
fastapi==0.95.0
pydantic==1.10.7
pymongo==4.3.3
pytest==7.2.2
uvicorn==0.21.1
python-socketio==5.8.0